refactor storybook taxonomic filter scene
